+++ src/main/java/com/easypost/model/Address.java
@@ -236,46 +236,6 @@ public final class Address extends BaseAddress {
     }
 
     /**
-     * Create Address object from parameter map
-     * and immediately verify it with a specific carrier.
-     *
-     * @param params  Map of address parameters.
-     * @param carrier Carrier to verify address with.
-     * @return Address object.
-     * @throws EasyPostException when the request fails.
-     */
-    public static Address createAndVerifyWithCarrier(final Map<String, Object> params, final String carrier)
-            throws EasyPostException {
-        return createAndVerifyWithCarrier(params, carrier, null);
-    }
-
-    /**
-     * Create Address object from parameter map
-     * and immediately verify it with a specific carrier.
-     *
-     * @param params  Map of address parameters.
-     * @param carrier Carrier to verify address with.
-     * @param apiKey  API key to use in request (overrides default API key).
-     * @return Address object.
-     * @throws EasyPostException when the request fails.
-     */
-    public static Address createAndVerifyWithCarrier(final Map<String, Object> params, final String carrier,
-                                                     final String apiKey) throws EasyPostException {
-        Map<String, Object> wrappedParams = new HashMap<String, Object>();
-        wrappedParams.put("address", params);
-        wrappedParams.put("carrier", carrier);
-
-        AddressVerifyResponse response;
-        response = request(RequestMethod.POST, String.format("%s/create_and_verify", classURL(Address.class)),
-                wrappedParams, AddressVerifyResponse.class, apiKey);
-
-        if (response.getMessage() != null) {
-            response.getAddress().setMessage(response.getMessage());
-        }
-        return response.getAddress();
-    }
-
-    /**
      * Verify this Address object.
      *
      * @return Address object.
@@ -298,43 +258,6 @@ public final class Address extends BaseAddress {
                 request(RequestMethod.GET, String.format("%s/verify", instanceURL(Address.class, this.getId())), null,
                         AddressVerifyResponse.class, apiKey);
 
-        if (response.getMessage() != null) {
-            response.getAddress().setMessage(response.getMessage());
-        }
-        return response.getAddress();
-    }
-
-    /**
-     * Verify this Address object with a specific carrier.
-     *
-     * @param carrier Carrier to verify address with.
-     * @return Address object.
-     * @throws EasyPostException when the request fails.
-     */
-    public Address verifyWithCarrier(final String carrier) throws EasyPostException {
-        return this.verifyWithCarrier(carrier, null);
-    }
-
-    /**
-     * Verify this Address object with a specific carrier.
-     *
-     * @param carrier Carrier to verify address with.
-     * @param apiKey  API key to use in request (overrides default API key).
-     * @return Address object.
-     * @throws EasyPostException when the request fails.
-     */
-    public Address verifyWithCarrier(final String carrier, final String apiKey) throws EasyPostException {
-        Map<String, Object> wrappedParams = new HashMap<String, Object>();
-        wrappedParams.put("carrier", carrier);
-
-        AddressVerifyResponse response;
-        response =
-                request(RequestMethod.GET, String.format("%s/verify", instanceURL(Address.class, this.getId())), null,
-                        AddressVerifyResponse.class, apiKey);
-
-        if (response.getMessage() != null) {
-            response.getAddress().setMessage(response.getMessage());
-        }
         return response.getAddress();
     }
 }